backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Scraping
tweepy==4.14.0
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2

# Background Jobs
celery==5.3.4
redis==5.0.1

# Notifications
python-telegram-bot==21.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0